Add attribution header to copied code.
